change to uint32

+++ options_block_based_table.go
@@ -176,8 +176,8 @@ func (opts *BlockBasedTableOptions) SetBlockCache(cache *Cache) {
 }
 
 // set cache size with C.rocksdb_block_based_options_set_cache_size
-func (opts *BlockBasedTableOptions) SetCacheSize(capacity uint64) {
-	C.rocksdb_block_based_options_set_cache_size(opts.c, C.uint64_t(capacity))
+func (opts *BlockBasedTableOptions) SetCacheSize(capacity uint32) {
+	C.rocksdb_block_based_options_set_cache_size(opts.c, C.uint32_t(capacity))
 }
 
 // SetBlockCacheCompressed sets the cache for compressed blocks.
